Make decorator a separate class to contain helper methods
Add `decorator` accessor at AR model and provide BaseDecorator for user to extend upon.
AR model can get an instance of decorator this way,
and then call the methods from there:

    record.decorator.foo_bar()

I feel extending the model itself can easily lead to leaky abstractions.
So I went the route to create a separate class for this.
It should also be beneficial for dev to be aware that the helper
method sits at decorator level not model level.